Geekyshows/python/16.Use Free LLM with LangChain Python/ch15/requirements.txt
langchain==0.3.21
python-decouple==3.8
huggingface-hub==0.30.2
